Add Longer

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] a = {100, 5, 26, 58, 50};
+        long[] a = {100, 5, 26, 58, 50};
 
         SalesManager manager = new SalesManager(a);
